+++ app/src/main/java/com/omarea/filter/LightHistory.kt
@@ -2,5 +2,5 @@ package com.omarea.filter
 
 class LightHistory {
     internal var time: Long = 0
-    internal var lux: Int = 0
+    internal var lux: Float = 0F
 }
